backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.37
alembic==1.14.0
pydantic==1.10.18
python-dotenv==1.0.1
pyyaml==6.0.2
